refactor: Price를 immutable한 wraper로 변경, Ranking 에 stream API 도입

+++ src/main/java/lotto/Main.java
@@ -23,7 +23,7 @@ public class Main {
     }
 
     private static List<LottoTicket> generateLottoTickets(Price income) {
-        int ticketNumber = income.floorDivide(new Price(1000));
+        long ticketNumber = income.floorDivide(new Price(1000));
         System.out.println(ticketNumber + "개를 구매했습니다.");
 
         List<LottoTicket> lottoTickets = new ArrayList<>();
